requirements.txt
aiogram==2.25.1
telethon==1.34.0
python-dotenv==1.0.1
aiohttp==3.9.3
